+++ OKX.Api/Account/Internal/OkxAccountFeeTypeContainer.cs
@@ -1,0 +1,13 @@
+﻿namespace OKX.Api.Account;
+
+/// <summary>
+/// OkxAccountOkxTypeContainer
+/// </summary>
+internal record OkxAccountFeeTypeContainer
+{
+    /// <summary>
+    /// Fee type
+    /// </summary>
+    [JsonProperty("feeType")]
+    public OkxAccountFeeType Payload { get; set; }
+}