+++ OKX.Api/Public/Responses/OkxPublicInstrument.cs
@@ -133,7 +133,7 @@ public record OkxPublicInstrument
     /// Only applicable to SPOT/MARGIN, return "" for all other business lines
     /// </summary>
     [JsonProperty("openType")]
-    public string? OpenType { get; set; } // TODO: Enumerate
+    public OkxPublicOpenType? OpenType { get; set; }
 
     /// <summary>
     /// Expiry timestamp